[FIX] id Long타입으로 변경

+++ src/main/java/sopt/org/CarrotServer/controller/sale/dto/response/SaleResponseDto.java
@@ -4,13 +4,11 @@ package sopt.org.CarrotServer.controller.sale.dto.response;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
-import sopt.org.CarrotServer.domain.sale.Sale;
-import sopt.org.CarrotServer.domain.sale.SaleStatus;
 
 @Getter
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 public class SaleResponseDto {
-    private Integer id;
+    private Long id;
     private String title;
     private String imgUrl;
     private String location;
@@ -21,9 +19,7 @@ public class SaleResponseDto {
     private Integer likeCount;
     private Boolean isCheckLike;
 
-
-
-    public static SaleResponseDto of(Integer id, String title, String imgUrl, String location, String time,
+    public static SaleResponseDto of(Long id, String title, String imgUrl, String location, String time,
                                      Boolean isUpdated, Integer price, Boolean isDiscount, Integer likeCount, Boolean isCheckLike) {
         return new SaleResponseDto(id, title, imgUrl, location, time, isUpdated, price, isDiscount, likeCount, isCheckLike);
     }
